--- consai_ros2/robocup_ssl_comm/src/robot_status_node.cpp
@@ -1,21 +1,0 @@
-// Copyright (c) 2024 ibis-ssl
-//
-// Use of this source code is governed by an MIT-style
-// license that can be found in the LICENSE file or at
-// https://opensource.org/licenses/MIT.
-
-#include <memory>
-
-#include "robocup_ssl_comm/robot_status_component.hpp"
-
-int main(int argc, char * argv[])
-{
-  rclcpp::init(argc, argv);
-  rclcpp::executors::SingleThreadedExecutor exe;
-  rclcpp::NodeOptions options;
-  auto node = std::make_shared<robocup_ssl_comm::GrSimRobotStatus>(options);
-  exe.add_node(node->get_node_base_interface());
-  exe.spin();
-  rclcpp::shutdown();
-  return 0;
-}
