+++ crane_robot_receiver/CMakeLists.txt
@@ -35,6 +35,14 @@ ament_auto_add_executable(diagnostic_publisher_node
 )
 add_backward(diagnostic_publisher_node)
 
+ament_auto_add_executable(grsim_robot_status_node
+        src/grsim_robot_status_node.cpp
+)
+# これがないと参照未解決エラーが出る
+target_link_libraries(grsim_robot_status_node
+        ${robocup_ssl_msgs_LIBRARIES})
+add_backward(grsim_robot_status_node)
+
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
   ament_lint_auto_find_test_dependencies()
